skill requirement helper method
