+++ src/tools/wiki/getQuestInfo.ts
@@ -11,6 +11,7 @@ import { findTemplates } from '../../utils/templateHelpers.js';
 import { QuestInfoToolResponse } from '../../zod';
 import { QuestInfoToolResponseType } from '../../types/osrsMcp.js';
 import { getItemRequirements, getQuestPreRequisites } from '../../core/quest';
+import { getSkillRequirements } from '../../core/quest/getSkillRequirements.js';
 
 export async function getQuestInfo(
 	questName: string,
@@ -44,44 +45,11 @@ export async function getQuestInfo(
 		response.questGiver = npcNameRaw.split('|')[0].trim()
 	}
 
-
 	if (requirements) {
 		response.itemReqs = getItemRequirements(requirements)
 		response.questReqs = getQuestPreRequisites(requirements)
+		response.skillReqs = getSkillRequirements(requirements)
 	}
-
-	// Parse skill requirements from the requirements string
-	const skillReqs: Record<string, number> = {}
-	if (requirements) {
-		// Extract SCP templates from the requirements string
-		const scpTemplates = extractTemplatesFromString(requirements)
-
-		// Filter for SCP templates (skill clickpic templates)
-		const skillTemplates = scpTemplates.filter(template =>
-			template.title === 'SCP' || template.title === 'Scp' || template.title === 'scp',
-		)
-
-		// Process each SCP template to extract skill and level
-		for (const template of skillTemplates) {
-			// SCP template typically has parameters:
-			// 1 = skill name (e.g., "Mining", "Smithing")
-			// 2 = skill level required (e.g., "50", "30")
-			const skillName = template.parameters['1']?.trim()
-			const levelStr = template.parameters['2']?.trim()
-
-			if (skillName && levelStr) {
-				const level = parseInt(levelStr, 10)
-				if (!isNaN(level)) {
-					// Store the skill requirement
-					// If the same skill appears multiple times, keep the highest requirement
-					if (!skillReqs[skillName] || skillReqs[skillName] < level) {
-						skillReqs[skillName] = level
-					}
-				}
-			}
-		}
-	}
-	response.skillReqs = skillReqs
 
 	// Parse recommended items from the recommended string
 	const recommendedItems: Record<string, number> = {}
